Upgrade web2py to 1.85.1

+++ gluon/sqlhtml.py
@@ -191,6 +191,49 @@ class OptionsWidget(FormWidget):
 
         return SELECT(*opts, **attr)
 
+class ListWidget(StringWidget):
+    @staticmethod
+    def widget(field,value,**attributes):
+        _id = '%s_%s' % (field._tablename, field.name)
+        _name = field.name
+        if field.type=='list:integer': _class = 'integer'
+        else: _class = 'string'
+        requires = field.requires
+        items=[LI(INPUT(_id=_id,_class=_class,_name=_name,value=v,hideerror=True)) \
+                   for v in value or ['']]
+        script=SCRIPT("""
+// from http://refactormycode.com/codes/694-expanding-input-list-using-jquery
+(function(){
+jQuery.fn.grow_input = function() {
+  return this.each(function() {
+    var ul = this;
+    jQuery(ul).find(":text").keypress(function (e) { return (e.which == 13) ? pe(ul) : true; });
+  });
+};
+function pe(ul) {
+  var new_line = ml(ul);
+  rel(ul);
+  new_line.appendTo(ul);
+  new_line.find(":text").focus();
+  return false;
+}
+function ml(ul) {
+  var line = jQuery(ul).find("li:first").clone(true);
+  line.find(':text').val('');
+  return line;
+}
+function rel(ul) {
+  jQuery(ul).find("li").each(function() {
+    var trimmed = jQuery.trim(jQuery(this.firstChild).val());
+    if (trimmed=='') jQuery(this).remove(); else jQuery(this.firstChild).val(trimmed);
+  });
+}
+})();
+jQuery(document).ready(function(){jQuery('#%s_grow_input').grow_input();});
+""" % _id)
+        attributes['_id']=_id+'_grow_input'
+        return TAG[''](UL(*items,**attributes),script)
+
 
 class MultipleOptionsWidget(OptionsWidget):
 
@@ -555,6 +598,7 @@ class SQLFORM(FORM):
         radio = RadioWidget,
         checkboxes = CheckboxesWidget,
         autocomplete = AutocompleteWidget,
+        list = ListWidget,
         ))
 
     FIELDNAME_REQUEST_DELETE = 'delete_this_record'
@@ -726,6 +770,8 @@ class SQLFORM(FORM):
                     inp = self.widgets.multiple.widget(field, default)
                 if fieldname in keepopts:
                     inpval = TAG[''](*inp.components)
+            elif str(field.type).startswith('list'):
+                inp = self.widgets.list.widget(field,default)
             elif field.type == 'text':
                 inp = self.widgets.text.widget(field, default)
             elif field.type == 'password':
@@ -907,9 +953,15 @@ class SQLFORM(FORM):
             )
 
         if not ret and self.record and self.errors:
+            ### if there are errors in update mode
+            # and some errors refers to an already uploaded file
+            # delete error if
+            # - user not trying to upload a new file
+            # - there is existing file and user is not trying to delete it
+            # this is because removing the file may not pass validation
             for key in self.errors.keys():
-                if not request_vars.get(key,None) \
-                        and self.table[key].type=='upload' \
+                if self.table[key].type == 'upload' \
+                        and request_vars.get(key,None) in (None,'') \
                         and self.record[key] \
                         and not key+UploadWidget.ID_DELETE_SUFFIX in request_vars:
                     del self.errors[key]
@@ -950,14 +1002,15 @@ class SQLFORM(FORM):
                                '%s != %s' % (record_id, self.record_id)
 
         if requested_delete:
-            if keyed:
-                qry = reduce(lambda x,y: x & y,
-                             [self.table[k]==record_id[k] for k in self.table._primarykey])
-                if self.table._db(qry).delete():
-                    self.vars.update(record_id)
-            else:
-                self.table._db(self.table.id == self.record.id).delete()
-                self.vars.id = self.record.id
+            if dbio:
+                if keyed:
+                    qry = reduce(lambda x,y: x & y,
+                                 [self.table[k]==record_id[k] for k in self.table._primarykey])
+                    if self.table._db(qry).delete():
+                        self.vars.update(record_id)
+                else:
+                    self.table._db(self.table.id == self.record.id).delete()            
+                    self.vars.id = self.record.id
             self.errors.clear()
             for component in self.elements('input, select, textarea'):
                 component['_disabled'] = True
@@ -990,6 +1043,7 @@ class SQLFORM(FORM):
                         fields[fieldname] = ''
                     else:
                         fields[fieldname] = self.record[fieldname]
+                    self.vars[fieldname] = fields[fieldname]
                     continue
                 elif hasattr(f,'file'):
                     (source_file, original_filename) = (f.file, f.filename)
